+++ recras-wordpress-plugin.php
@@ -6,7 +6,7 @@ Description: Easily integrate your Recras data into your own site
 Author: Recras
 Text Domain: recras-wp
 Domain Path: /lang
-Version: 2.2.2
+Version: 2.3.0
 
 Author URI: https://www.recras.nl/
 */